# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Photo Share API
  version: 0.1.0
  description: Initial API contract for the Photo Share MVP.
servers:
  - url: https://localhost:8080
    description: Local backend server
...
paths:
...
  /posts:
    get:
      tags:
        - posts
      summary: Get all posts
      operationId: listPosts
      responses:
        '200':
          description: A list of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'

    post:
      tags:
        - posts
      summary: Create a new post
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreatePostRequest:
      type: object
      required:
        - image_path
        - caption
      properties:
        image_path:
          type: string
          example: /demo/test.jpg
        caption:
          type: string
          example: Sunset at the beach
        location:
          type: string
          nullable: true
          example: Kamakura
        camera_body:
          type: string
          nullable: true
          example: Sony A7 IV
        lens:
          type: string
          nullable: true
          example: FE 24-70mm F2.8 GM
    Post:
      type: object
      required:
        - id
        - user_id
        - image_path
        - caption
        - created_at
        - updated_at
        - author
      properties:
        id:
          type: integer
          format: int64
          example: 100
        user_id:
          type: integer
          format: int64
          example: 1
        image_path:
          type: string
          example: /demo/test.jpg
        caption:
          type: string
          example: Sunset at the beach
        location:
          type: string
          nullable: true
          example: Kamakura
        camera_body:
          type: string
          nullable: true
          example: Sony A7 IV
        lens:
          type: string
          nullable: true
          example: FE 24-70mm F2.8 GM
        created_at:
          type: string
          format: date-time
          example: 2026-04-14T12:00:00Z
        updated_at:
          type: string
          format: date-time
          example: 2026-04-14T12:00:00Z
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: invalid_input
        message:
          type: string
          example: Caption is required
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'